Removed `internal` module, replaced with `primitives` facade
* `Iter` is now a struct in a separate namespace, it is no longer
  exposed.
* Changed all occurrences of "Internal" used to describe something which
  is used to construct fundamental parsers to "Primitive".
* Moved `State` and `IntoInner` into private module parse_result, still exposed
  through the `primitives` facade.
